# consult-org-*: show inherited tags