@minad 2023-03-14
Great! Would be good too benchmark before and after against @jdtsmith's

https://github.com/jdtsmith/shakespeare.org

Unfortunately it seems the file doesn't use tags?