@scvalex 2023-03-14
# Results over a combined 10 runs:

| File | Old (s) | New (s) | Percentage increase |
| --- | ---: | ---: | ---: |
| `shakespeare.org` | 105 | 124 | 18% |
| `shakespeare.org` with tags | 143 | 169 | 18% |



# Testing methodology:

I ran the following on `shakespeare.org`:

```
(benchmark-run-compiled 10
  (let ((headings (consult-org--headings nil nil 'file)))
    (length headings)))
```

Old:  105s   (raw result: `(105.75758244699999 363 71.826926945)`)
New: 124s (raw result: `(124.460416937 344 83.507605352)`)

I added tags to the first four levels of headings with this sed command: `sed -i -E -e 's#^\* (.*)#\* \1     :first:#' -e 's#^\*\* (.*)#\*\* \1     :second:#' -e 's#^\*\*\* (.*)#\*\*\* \1     :third:#' -e 's#^\*\*\*\* (.*)#\*\*\*\* \1     :fourth:#' shakespeare.org`.

Re-running the `benchmark-run-compiled`, I got:

Old: 143s (raw result: `(143.03177621700002 274 108.70956182399999)`)
New: 169s (raw result: `(169.44713967799998 314 124.14569273600003)`)