@scvalex 2023-03-14
It does get slower by adding more tags (third row in new): 

| File | Old time (s) | New time (s) | Percentage increase |
| --- | ---: | ---: | ---: |
| `shakespeare.org` | 105 | 124 | 18% |
| `shakespeare.org` with 1 tag per heading | 143 | 169 | 18% |
| `shakespeare.org` with 3 tags per heading | 154 | 230 | 49% |

Raw results:
Old: `(154.736029183 345 114.31684586600005)`
New: `(230.903019271 373 177.70347836099995)`

Sed command:
```
sed -i -E -e 's#^\* (.*)#\* \1     :first:one:unu:#' -e 's#^\*\* (.*)#\*\* \1     :second:two:doi:#' -e 's#^\*\*\* (.*)#\*\*\* \1     :third:three:trei:#' -e 's#^\*\*\*\* (.*)#\*\*\*\* \1     :fourth:four:patru:#' shakespeare.org
```