@minad 2023-03-14
On 3/14/23 14:24, Alexandru Scvorțov wrote:
> It does get slower by adding more tags (third row in new):
> 
> File 	Old time (s) 	New time (s) 	Percentage increase
> |shakespeare.org| 	105 	124 	18%
> |shakespeare.org| with 1 tag per heading 	143 	169 	18%
> |shakespeare.org| with 3 tags per heading 	154 	230 	49%
> 
> Raw results:
> Old: |(154.736029183 345 114.31684586600005)|
> New: |(230.903019271 373 177.70347836099995)|

Hmm, a 50% slow down is too much. Could you please take a profile to
determine the hotspot within the org-get-tags function?

We could enable the slow org-get-tags call only if
org-use-tag-inheritance is non-nil and otherwise use the tags value from
the heading components.