@minad 2023-03-14
@jdtsmith

> Good idea. I added tags to shakespeare.org for all titles, which includes :PLAY, :ACT, :SCENE, etc. Most every heading will inherit tags.

Ideally could you provide multiple variants? With or without tags, with different levels of nesting, with or without whitespace indentation, and maybe more variations. I think this would be quite useful for issues like this one and other benchmark purposes.