@jdtsmith 2023-03-14
I added a `_notags` and `_depthN`.   

Not sure about the indentation idea.  `org-indent` is as you know "fake" indentation.  Only lists and blocks does org actually indent I think. I could create a variant where the speakers in a given scene form (a giant) definition list, but it would not be very nested.