@jdtsmith 2023-03-14
BTW, 10s is a pretty long time to wait for headings... building the entire file from XML source takes only 0.6s on a slow machine with python.