@scvalex 2023-03-14
It's 10s for 10 runs.