@scvalex 2023-03-15
@minad Sorry, but I think I'm out of this PR.

I agree with you that 50% slowdown is not good for a general purpose library, but it happens to be acceptable for me.  To give concrete numbers, my notes for the last 3 years are only 5% of the size of shakespeare.org, so I don't even notice the slowdown and I will probably never write enough to ever notice it.

As you said, the next step here is profiling `org-get-tags` and either making that faster, or writing a new function that takes advantage of the fact that we're visiting all the headlines in a file in order.  That's more elisp than I am willing to learn in order to get the hacked function upstreamed.

> We could enable the slow org-get-tags call only if
> org-use-tag-inheritance is non-nil and otherwise use the tags value from
> the heading components.

Tag inheritance is enabled by default, so using that as the switch for the slow logic wouldn't really help anyone.  At best, it'd be an escape hatch in case it broke somebody's workflow.