@jdtsmith 2023-03-15
@yantar92 could probably tell you instantly what the hotspots in `org-get-tags` are.  I also wouldn't be surprised if the new `org-element-cache` improvements the devs have been working on will improve speed in that function for large/nested/tag-rich files.   @scvalex are you on org 9.6?