@yantar92 2023-03-15
JD Smith ***@***.***> writes:

> @yantar92 could probably tell you instantly what the hotspots in
> `org-get-tags` are.

Try to benchmark with `org-element--cache-self-verify' set to nil.
Cache is in debug mode in Org 9.6 release.

> I also wouldn't be surprised if the new `org-element-cache` improvements the devs have been working on will improve speed in that function for large/nested/tag-rich files.   @scvalex are you on org 9.6? 

Having org-element-cache in `org-get-tags` implies Org 9.6 or main branch.

-- 
Ihor Radchenko // yantar92,
Org mode contributor,
Learn more about Org mode at <https://orgmode.org/>.
Support Org development at <https://liberapay.com/org-mode>,
or support my work at <https://liberapay.com/yantar92>