@scvalex 2023-03-15
I am on org 9.6.1 and Emacs 30.0.50 (the tip of master from a few weeks ago with native compilation enabled).

The new results are the bottom 2 rows:

| File | `cache-self-verify` | Old time (s) | New time (s) | Percentage increase |
| --- | --- | ---: | ---: | ---: |
| old `sp.org` | `'t` | 105 | 124 | 18% |
| old `sp.org` with 1 tag per heading | `'t` | 143 | 169 | 18% |
| old `sp.org` with 3 tags per heading | `'t` | 154 | 230 | 49% |
| old `sp.org` with 3 tags per heading | `'nil` | 173 | 201 | 16% |
| new `sp.org` | `'nil` | 121 | 127 | 5% |
| new `sp_depth3.org` | `'nil` | 4.64 | 4.82 | 6% |