@scvalex 2023-03-15
Oh, and I tested with this code:

(setq org-element--cache-self-verify nil)
(benchmark-run-compiled 10
  (let ((headings (consult-org--headings nil nil 'file)))
    (length headings)))

Raw results:

3tags Old: (173.813383024 308 138.12313946099994)
3tags New: (201.027180562 319 151.90207473400005)

new sp.org Old: `(121.90266349000001 269 73.33232696399989)`
new sp.org New: `(127.368035139 298 74.67290736900009)`

sp_depth3.org Old: `(4.645136953 11 1.8673587159998988)`
sp_depth3.org New: `(4.823771954000001 12 2.011230398999942)`