@scvalex 2023-03-15
That's a bit annoying to do.  What would that determine?