@yantar92 2023-03-15
Well. You don't have to do it yourself. I am mostly curious. Because if the problem is with emacs#58558, not much can be done on lisp side.