@minad 2023-03-15
Merged, a 5%-10% slowdown is perfectly acceptable for this feature.