@scvalex 2023-03-15
Thank you, but my interpretation of this thread is that it's going to be perhaps 50% slowdown in the near future, and then decrease to 10% when further org development happens.