@minad 2023-03-15
Yes, but that's good enough. I also added the `org-use-tag-inheritance` escape hatch. I already use Emacs 29.1 Org 9.6.1 and I have set `org-element--cache-self-verify` to nil in my config. As far as I have understood, the Org self verification mechanism will be disabled soon by default.