@jdtsmith 2023-03-15
I think the new Shakespeare.org has a more sensible/typical tag arrangement, 1000 vs. 100,000 tags in a (large) file.  But I don't use tags too often so I may be biased.